feat: Add a flash and sound notification when a core timer completes
